Implement sorting at ElasticSearch Query
Sending sorting fields are optional.
A sort object should be send, which can include multiple fields to sort by, and their order direction to sort by (ascending/descending).
A mimimum score will be used of 0.5, or another minumum treshold can be passed as a parameter.
If no sort object is passes the results will be sorted as before by the score.