Added Download Endpoint for Downloading CSV/XLS/XLSX Files 171ec64
Added download endpoint so that the ElasticSearch results can be downloaded in a file.
Additional parameters include file_format, file_name and column_mapping.

With the column mapping one can define what the column headers should be in the file, for example, change 'city' to 'עיר'
When the column_mapping is sent, only the fields mentioned are included, so this is also a way to leave certain fields out if not needed